ODIN-286 added null detection for proxyHost/port
In the event that we have the config proxyHost=${proxy.host} and
proxyPort=${proxy.port}, then these variables may have an empty string
substitution in some environments; and an actual value in others.
Added blank protection into the client configuration builder so that
we handle this use case for proxyHost/port only as we know that the
default behaviour within AWS is for no proxy.