INTERLOK-2618 add force-path-style-access to S3
- Due to VPN / Umbrella nonsense, http://127.0.0.1:4572 doesn't
appear to work; http://localhost:4572 does, but causes you to use
virtual host names when you attempt to access S3. Add a specific
flag to "force path style access" in the builder.
- Remove validation restrictions on CustomEndpoint; instead we
want to track behaviour at runtime so that if we have "blank
configuration" we default back to Region style configuration.