deps: switch to using aws-sdk-bom
Using the BOM means we don't need to care about the transitive
dependencies from amazon-kinesis-client that refer to outdated version.

Dependencies left within each sub-project since there is a PR that
includes AWS-SDK-V2 so a hard dependency in the parent for AWS-SDK-V1
is probably a bad thing.

Switch netty to use the bom as well.